- harmonisation des types
(Logical change 1.1220)


git-svn-id: https://scm.gforge.inria.fr/authscm/jmcoutey/svn/pm2/trunk@11455 ee107e0b-c2fa-0310-842a-ec2723fb34f4

+++ pm2/pm2/ntbx/include/ntbx_types.h
@@ -105,9 +105,9 @@ typedef struct s_ntbx_client
   p_tbx_slist_t        remote_alias;
   tbx_bool_t           read_ok;
   tbx_bool_t           write_ok;
-  unsigned long        read_rq;
+  size_t	       read_rq;
   tbx_bool_t           read_rq_flag;
-  unsigned long        write_rq;
+  size_t               write_rq;
   tbx_bool_t           write_rq_flag;
   void                *specific;
 } ntbx_client_t;
